backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
celery>=5.3.0
redis>=5.0.0
firebase-admin>=6.4.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
gunicorn>=21.2.0
python-dotenv>=1.0.1
pytz>=2024.1
drf-spectacular>=0.27.1
django-cors-headers>=4.3.0
whitenoise>=6.6.0
pytest>=8.0.0
pytest-django>=4.8.0
